Mount app at two paths, one handling requests on IO threads
This requires a slightly older version of Undertow than what Immutant
drags in, but it is the same version used by the Undertow benchmark in
this repo.

My assumption is that benchmark_config.json is what's used to determine
the proper URI's for each test, and therefore only those that should NOT
be dispatched are prefixed with the "/io" path.